Remove `include/grpc/impl/codegen/byte_buffer.h`
